# backend/iotd/api/routes/auth.py
"""Auth routes — signup, login, refresh, /me."""

from typing import Annotated

from fastapi import APIRouter, Depends, status

from iotd.api.cache import NO_STORE
from iotd.api.controllers.auth import AuthController
from iotd.api.schemas.auth import LoginRequest, RefreshRequest, SignupRequest, TokenPair, UserPublic
from iotd.dependencies import get_auth_controller, get_current_user
from iotd.models.user import User

router = APIRouter(prefix="/auth", tags=["auth"])


@router.post(
    "/signup",
    response_model=TokenPair,
    response_model_by_alias=True,
    status_code=status.HTTP_201_CREATED,
    summary="Create a new account and receive a token pair",
    dependencies=[NO_STORE],
)
async def signup(
    body: SignupRequest,
    controller: Annotated[AuthController, Depends(get_auth_controller)],
) -> TokenPair:
    return await controller.signup(body)


@router.post(
    "/login",
    response_model=TokenPair,
    response_model_by_alias=True,
    summary="Exchange email + password for a token pair",
    dependencies=[NO_STORE],
)
async def login(
    body: LoginRequest,
    controller: Annotated[AuthController, Depends(get_auth_controller)],
) -> TokenPair:
    return await controller.login(body)


@router.post(
    "/refresh",
    response_model=TokenPair,
    response_model_by_alias=True,
    summary="Exchange a refresh token for a fresh pair",
    dependencies=[NO_STORE],
)
async def refresh(
    body: RefreshRequest,
    controller: Annotated[AuthController, Depends(get_auth_controller)],
) -> TokenPair:
    return await controller.refresh(body.refresh_token)


@router.get(
    "/me",
    response_model=UserPublic,
    response_model_by_alias=True,
    summary="Return the authenticated user's profile",
    dependencies=[NO_STORE],
)
async def me(user: Annotated[User, Depends(get_current_user)]) -> UserPublic:
    return UserPublic.model_validate(user)