# web/openapi.json
{"openapi": "3.1.0", "info": {"title": "Islamic On This Day", "description": "Read-only API serving one verified Islamic-history event per day, in both Hijri and Gregorian calendars. Backed by the data-pipeline's curated SQLite database.", "version": "0.1.0"}, "paths": {"/health": {"get": {"tags": ["system"], "summary": "Liveness + DB ping", "description": "Return the API health snapshot.", "operationId": "health_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HealthResponse"}}}}}}}, "/api/v1/today": {"get": {"tags": ["today"], "summary": "Headline event, secondary rails, observance for the current day", "description": "Return the Today payload for the current UTC day.\n\nA CDN can serve the same payload to thousands of readers in the same\ntime zone with one origin hit; the dependency stamps a Cache-Control\nthat expires at the next UTC midnight so the rollover is automatic.", "operationId": "get_today_api_v1_today_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TodayResponse"}}}}}}}, "/api/v1/events/{slug}": {"get": {"tags": ["events"], "summary": "Single-event detail by slug", "description": "Return the full event payload for ``slug``.", "operationId": "get_event_api_v1_events__slug__get", "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 160, "pattern": "^[a-z0-9][a-z0-9\\-]*$", "title": "Slug"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EventDetail"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/lessons": {"get": {"tags": ["lessons"], "summary": "List lessons with optional category filter", "description": "Paginated lesson list.", "operationId": "list_lessons_api_v1_lessons_get", "parameters": [{"name": "category", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string", "maxLength": 48}, {"type": "null"}], "description": "Filter by lesson category.", "title": "Category"}, "description": "Filter by lesson category."}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "default": 30, "title": "Limit"}}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Offset"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LessonListResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/lessons/{slug}": {"get": {"tags": ["lessons"], "summary": "Single lesson detail by slug", "description": "Return the full lesson payload for ``slug``.", "operationId": "get_lesson_api_v1_lessons__slug__get", "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 160, "pattern": "^[a-z0-9][a-z0-9\\-]*$", "title": "Slug"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LessonDetail"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/observances": {"get": {"tags": ["observances"], "summary": "List every recurring annual observance", "description": "Return every observance, ordered by Hijri month + day.", "operationId": "list_observances_api_v1_observances_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ObservanceDetail"}, "type": "array", "title": "Response List Observances Api V1 Observances Get"}}}}}}}, "/api/v1/observances/{slug}": {"get": {"tags": ["observances"], "summary": "Single observance by slug", "description": "Look up one observance.", "operationId": "get_observance_api_v1_observances__slug__get", "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[a-z0-9][a-z0-9\\-]*$", "title": "Slug"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ObservanceDetail"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/people/{slug}": {"get": {"tags": ["people"], "summary": "Single-person detail by slug", "description": "Return the full person payload for ``slug``.", "operationId": "get_person_api_v1_people__slug__get", "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[a-z0-9][a-z0-9\\-]*$", "title": "Slug"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PersonDetail"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/recent": {"get": {"tags": ["recent"], "summary": "Recent days \u2014 headline and observance for the last 7 calendar days", "description": "Return the Recent payload \u2014 the catch-up window.", "operationId": "get_recent_api_v1_recent_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RecentResponse"}}}}}}}, "/api/v1/auth/signup": {"post": {"tags": ["auth"], "summary": "Create a new account and receive a token pair", "operationId": "signup_api_v1_auth_signup_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SignupRequest"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenPair"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/login": {"post": {"tags": ["auth"], "summary": "Exchange email + password for a token pair", "operationId": "login_api_v1_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenPair"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/refresh": {"post": {"tags": ["auth"], "summary": "Exchange a refresh token for a fresh pair", "operationId": "refresh_api_v1_auth_refresh_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RefreshRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenPair"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/me": {"get": {"tags": ["auth"], "summary": "Return the authenticated user's profile", "operationId": "me_api_v1_auth_me_get", "parameters": [{"name": "Authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserPublic"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/bookmarks": {"get": {"tags": ["bookmarks"], "summary": "List the authenticated user's saved items", "operationId": "list_bookmarks_api_v1_bookmarks_get", "parameters": [{"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 200, "minimum": 1, "default": 100, "title": "Limit"}}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Offset"}}, {"name": "Authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BookmarkList"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["bookmarks"], "summary": "Save a new bookmark", "operationId": "create_bookmark_api_v1_bookmarks_post", "parameters": [{"name": "Authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BookmarkCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BookmarkOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/bookmarks/{bookmark_id}": {"delete": {"tags": ["bookmarks"], "summary": "Delete one of the authenticated user's bookmarks", "operationId": "delete_bookmark_api_v1_bookmarks__bookmark_id__delete", "parameters": [{"name": "bookmark_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Bookmark Id"}}, {"name": "Authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"BookmarkCreate": {"properties": {"targetKind": {"type": "string", "enum": ["event", "lesson", "observance", "person"], "title": "Targetkind"}, "targetSlug": {"type": "string", "maxLength": 160, "minLength": 1, "pattern": "^[a-z0-9][a-z0-9\\-]*$", "title": "Targetslug"}, "note": {"anyOf": [{"type": "string", "maxLength": 512}, {"type": "null"}], "title": "Note"}}, "additionalProperties": false, "type": "object", "required": ["targetKind", "targetSlug"], "title": "BookmarkCreate", "description": "Inputs for ``POST /api/v1/bookmarks``."}, "BookmarkList": {"properties": {"items": {"items": {"$ref": "#/components/schemas/BookmarkOut"}, "type": "array", "title": "Items"}, "total": {"type": "integer", "title": "Total"}}, "additionalProperties": false, "type": "object", "required": ["items", "total"], "title": "BookmarkList", "description": "Paginated bookmarks payload."}, "BookmarkOut": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "targetKind": {"type": "string", "enum": ["event", "lesson", "observance", "person"], "title": "Targetkind"}, "targetSlug": {"type": "string", "title": "Targetslug"}, "targetTitle": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Targettitle"}, "targetTitleAr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Targettitlear"}, "targetTitleFr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Targettitlefr"}, "note": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Note"}, "createdAt": {"type": "string", "format": "date-time", "title": "Createdat"}}, "additionalProperties": false, "type": "object", "required": ["id", "targetKind", "targetSlug", "targetTitle", "targetTitleAr", "targetTitleFr", "note", "createdAt"], "title": "BookmarkOut", "description": "A single bookmark \u2014 what the saves list returns."}, "DatasetSnapshot": {"properties": {"eventCount": {"type": "integer", "title": "Eventcount"}, "lessonCount": {"type": "integer", "title": "Lessoncount"}, "observanceCount": {"type": "integer", "title": "Observancecount"}, "personCount": {"type": "integer", "title": "Personcount"}, "builtAt": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Builtat"}, "ageHours": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Agehours"}}, "additionalProperties": false, "type": "object", "required": ["eventCount", "lessonCount", "observanceCount", "personCount"], "title": "DatasetSnapshot", "description": "Per-resource row counts + dataset freshness signal.\n\n``built_at`` and ``age_hours`` come from the most recent ``updated_at``\non the ``events`` table \u2014 that's the signal that a pipeline rebuild\nactually happened, even when no rows were added or removed."}, "DisputedPosition": {"properties": {"rank": {"type": "integer", "title": "Rank"}, "value": {"type": "string", "title": "Value"}, "scholars": {"type": "string", "title": "Scholars"}, "weight": {"type": "string", "enum": ["primary", "notable", "minority"], "title": "Weight"}}, "additionalProperties": false, "type": "object", "required": ["rank", "value", "scholars", "weight"], "title": "DisputedPosition", "description": "One scholarly position on a disputed date or fact.\n\n``weight`` is a discriminant (``primary | notable | minority``) \u2014 the\nfront-end maps it to a localised label."}, "EventDetail": {"properties": {"id": {"type": "string", "title": "Id"}, "title": {"type": "string", "title": "Title"}, "titleAr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Titlear"}, "titleFr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Titlefr"}, "era": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Era"}, "importance": {"type": "string", "title": "Importance"}, "verificationStatus": {"type": "string", "enum": ["scholar_reviewed", "cross_verified", "single_source", "unverified"], "title": "Verificationstatus"}, "gregorian": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Gregorian"}, "hijri": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Hijri"}, "location": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Location"}, "placeholder": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Placeholder"}, "noImage": {"type": "boolean", "title": "Noimage", "default": false}, "imageUrl": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Imageurl"}, "summary": {"type": "string", "title": "Summary"}, "summaryAr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Summaryar"}, "summaryFr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Summaryfr"}, "body": {"items": {"type": "string"}, "type": "array", "title": "Body", "default": []}, "bodyAr": {"items": {"type": "string"}, "type": "array", "title": "Bodyar", "default": []}, "bodyFr": {"items": {"type": "string"}, "type": "array", "title": "Bodyfr", "default": []}, "people": {"items": {"$ref": "#/components/schemas/PersonRef"}, "type": "array", "title": "People", "default": []}, "sources": {"items": {"$ref": "#/components/schemas/SourceRef"}, "type": "array", "title": "Sources", "default": []}, "disputed": {"type": "boolean", "title": "Disputed", "default": false}, "disputeAbout": {"anyOf": [{"type": "string", "enum": ["date", "detail", "interpretation"]}, {"type": "null"}], "title": "Disputeabout"}, "disputedPositions": {"items": {"$ref": "#/components/schemas/DisputedPosition"}, "type": "array", "title": "Disputedpositions", "default": []}, "sourceUrl": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Sourceurl"}}, "additionalProperties": false, "type": "object", "required": ["id", "title", "importance", "verificationStatus", "summary"], "title": "EventDetail", "description": "Full event projection used by the headline + event-detail surfaces.\n\nTrilingual: ``title`` (English) + ``title_ar`` (Arabic) + ``title_fr``\n(French) all surfaced; ``summary`` and ``body`` likewise. Front-end\nchooses which to render."}, "EventSummary": {"properties": {"id": {"type": "string", "title": "Id"}, "title": {"type": "string", "title": "Title"}, "titleAr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Titlear"}, "titleFr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Titlefr"}, "hijri": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Hijri"}, "gregorian": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Gregorian"}, "era": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Era"}, "importance": {"type": "string", "title": "Importance"}, "verificationStatus": {"type": "string", "enum": ["scholar_reviewed", "cross_verified", "single_source", "unverified"], "title": "Verificationstatus"}, "disputed": {"type": "boolean", "title": "Disputed", "default": false}, "disputeAbout": {"anyOf": [{"type": "string", "enum": ["date", "detail", "interpretation"]}, {"type": "null"}], "title": "Disputeabout"}}, "additionalProperties": false, "type": "object", "required": ["id", "title", "importance", "verificationStatus"], "title": "EventSummary", "description": "Slim event projection for \"On this day\" rotation rails \u2014 trilingual."}, "GregorianDate": {"properties": {"day": {"type": "integer", "title": "Day"}, "month": {"type": "string", "title": "Month"}, "year": {"type": "integer", "title": "Year"}, "weekday": {"type": "string", "title": "Weekday"}}, "additionalProperties": false, "type": "object", "required": ["day", "month", "year", "weekday"], "title": "GregorianDate", "description": "Calendar slot \u2014 Gregorian side."}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "HealthResponse": {"properties": {"status": {"type": "string", "title": "Status"}, "database": {"type": "string", "title": "Database"}, "version": {"type": "string", "title": "Version"}, "dataset": {"anyOf": [{"$ref": "#/components/schemas/DatasetSnapshot"}, {"type": "null"}]}}, "additionalProperties": false, "type": "object", "required": ["status", "database", "version"], "title": "HealthResponse", "description": "Liveness + DB connectivity probe + dataset snapshot."}, "HijriDate": {"properties": {"day": {"type": "integer", "title": "Day"}, "month": {"type": "string", "title": "Month"}, "monthShort": {"type": "string", "title": "Monthshort"}, "year": {"type": "integer", "title": "Year"}, "weekday": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Weekday"}}, "additionalProperties": false, "type": "object", "required": ["day", "month", "monthShort", "year"], "title": "HijriDate", "description": "Calendar slot \u2014 Hijri side."}, "LessonDetail": {"properties": {"kind": {"type": "string", "const": "lesson", "title": "Kind", "default": "lesson"}, "id": {"type": "string", "title": "Id"}, "title": {"type": "string", "title": "Title"}, "titleAr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Titlear"}, "titleFr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Titlefr"}, "category": {"type": "string", "title": "Category"}, "reference": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Reference"}, "summary": {"type": "string", "title": "Summary"}, "summaryAr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Summaryar"}, "summaryFr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Summaryfr"}, "body": {"items": {"type": "string"}, "type": "array", "title": "Body", "default": []}, "bodyAr": {"items": {"type": "string"}, "type": "array", "title": "Bodyar", "default": []}, "bodyFr": {"items": {"type": "string"}, "type": "array", "title": "Bodyfr", "default": []}, "quranRefs": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Quranrefs"}, "hadithRefs": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Hadithrefs"}, "sourceUrl": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Sourceurl"}, "sourceNotes": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Sourcenotes"}, "sourceNotesAr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Sourcenotesar"}, "sourceNotesFr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Sourcenotesfr"}}, "additionalProperties": false, "type": "object", "required": ["id", "title", "category", "summary"], "title": "LessonDetail", "description": "Full lesson projection used by the headline + lesson-detail surfaces.\n\nTrilingual: ``title`` (English) + ``title_ar`` (Arabic) + ``title_fr``\n(French) all surfaced; ``summary`` and ``body`` likewise. Front-end\nchooses which to render."}, "LessonListResponse": {"properties": {"items": {"items": {"$ref": "#/components/schemas/LessonSummary"}, "type": "array", "title": "Items", "default": []}, "total": {"type": "integer", "title": "Total"}, "limit": {"type": "integer", "title": "Limit"}, "offset": {"type": "integer", "title": "Offset"}}, "additionalProperties": false, "type": "object", "required": ["total", "limit", "offset"], "title": "LessonListResponse", "description": "Paginated list of lessons."}, "LessonSummary": {"properties": {"kind": {"type": "string", "const": "lesson", "title": "Kind", "default": "lesson"}, "id": {"type": "string", "title": "Id"}, "title": {"type": "string", "title": "Title"}, "titleAr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Titlear"}, "titleFr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Titlefr"}, "category": {"type": "string", "title": "Category"}, "reference": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Reference"}}, "additionalProperties": false, "type": "object", "required": ["id", "title", "category"], "title": "LessonSummary", "description": "Slim lesson projection for rotation rails \u2014 trilingual."}, "LoginRequest": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "password": {"type": "string", "maxLength": 128, "minLength": 1, "title": "Password"}}, "additionalProperties": false, "type": "object", "required": ["email", "password"], "title": "LoginRequest", "description": "Inputs for ``POST /api/v1/auth/login``."}, "ObservanceDetail": {"properties": {"id": {"type": "string", "title": "Id"}, "nameEn": {"type": "string", "title": "Nameen"}, "nameAr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Namear"}, "nameFr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Namefr"}, "descriptionEn": {"type": "string", "title": "Descriptionen"}, "descriptionAr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Descriptionar"}, "descriptionFr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Descriptionfr"}, "hijriMonth": {"type": "integer", "title": "Hijrimonth"}, "hijriDay": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Hijriday"}, "windowDays": {"type": "integer", "title": "Windowdays", "default": 1}, "importance": {"type": "string", "title": "Importance"}, "quranRefs": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Quranrefs"}, "hadithRefs": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Hadithrefs"}}, "additionalProperties": false, "type": "object", "required": ["id", "nameEn", "descriptionEn", "hijriMonth", "importance"], "title": "ObservanceDetail", "description": "A recurring Hijri-anchored Islamic observance (Eid, Mawlid, Arafah, \u2026).\n\nTrilingual: ``name``/``name_ar``/``name_fr`` and\n``description``/``description_ar``/``description_fr``."}, "ObservanceRef": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "nameAr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Namear"}, "nameFr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Namefr"}, "hijriDate": {"type": "string", "title": "Hijridate"}, "summary": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Summary"}, "summaryAr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Summaryar"}, "summaryFr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Summaryfr"}}, "additionalProperties": false, "type": "object", "required": ["id", "name", "hijriDate"], "title": "ObservanceRef", "description": "Active annual observance, if one is in season."}, "PersonDetail": {"properties": {"id": {"type": "string", "title": "Id"}, "fullNameEn": {"type": "string", "title": "Fullnameen"}, "fullNameAr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Fullnamear"}, "kunya": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Kunya"}, "laqab": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Laqab"}, "nisba": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Nisba"}, "role": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Role"}, "biography": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Biography"}, "isProphet": {"type": "boolean", "title": "Isprophet", "default": false}, "isSahabi": {"type": "boolean", "title": "Issahabi", "default": false}, "isAhlAlBayt": {"type": "boolean", "title": "Isahlalbayt", "default": false}, "imageUrl": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Imageurl"}, "imageBlockedReason": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Imageblockedreason"}, "wikidataQid": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Wikidataqid"}}, "additionalProperties": false, "type": "object", "required": ["id", "fullNameEn"], "title": "PersonDetail", "description": "Full person profile.\n\nThe image policy is enforced upstream by the pipeline: ``image_url`` is\nguaranteed to be ``None`` for any prophet, Sahabi, or member of Ahl\nal-Bayt. The frontend should not rely on its own check, but the value\nis also surfaced explicitly in ``image_blocked_reason`` when applicable."}, "PersonRef": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "nameAr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Namear"}, "nameFr": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Namefr"}, "role": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Role"}}, "additionalProperties": false, "type": "object", "required": ["id", "name"], "title": "PersonRef", "description": "A person attached to an event."}, "RecentDay": {"properties": {"date": {"type": "string", "title": "Date"}, "calendar": {"$ref": "#/components/schemas/TodayCalendar"}, "headline": {"anyOf": [{"$ref": "#/components/schemas/EventDetail"}, {"$ref": "#/components/schemas/LessonDetail"}, {"type": "null"}], "title": "Headline"}, "observance": {"anyOf": [{"$ref": "#/components/schemas/ObservanceRef"}, {"type": "null"}]}}, "additionalProperties": false, "type": "object", "required": ["date", "calendar"], "title": "RecentDay", "description": "One day inside the ``/api/v1/recent`` response."}, "RecentResponse": {"properties": {"days": {"items": {"$ref": "#/components/schemas/RecentDay"}, "type": "array", "title": "Days"}}, "additionalProperties": false, "type": "object", "required": ["days"], "title": "RecentResponse", "description": "The full ``/api/v1/recent`` payload."}, "RefreshRequest": {"properties": {"refreshToken": {"type": "string", "title": "Refreshtoken"}}, "additionalProperties": false, "type": "object", "required": ["refreshToken"], "title": "RefreshRequest", "description": "Inputs for ``POST /api/v1/auth/refresh``."}, "SignupRequest": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "password": {"type": "string", "maxLength": 128, "minLength": 8, "title": "Password"}, "displayName": {"anyOf": [{"type": "string", "maxLength": 64}, {"type": "null"}], "title": "Displayname"}}, "additionalProperties": false, "type": "object", "required": ["email", "password"], "title": "SignupRequest", "description": "Inputs for ``POST /api/v1/auth/signup``."}, "SourceRef": {"properties": {"label": {"type": "string", "title": "Label"}, "kind": {"type": "string", "enum": ["classical", "primary", "modern"], "title": "Kind"}, "verify": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Verify"}}, "additionalProperties": false, "type": "object", "required": ["label", "kind"], "title": "SourceRef", "description": "A citable source attached to an event."}, "TodayCalendar": {"properties": {"gregorian": {"$ref": "#/components/schemas/GregorianDate"}, "hijri": {"$ref": "#/components/schemas/HijriDate"}}, "additionalProperties": false, "type": "object", "required": ["gregorian", "hijri"], "title": "TodayCalendar", "description": "Co-primary calendar pair surfaced in the masthead."}, "TodayResponse": {"properties": {"today": {"$ref": "#/components/schemas/TodayCalendar"}, "headline": {"anyOf": [{"$ref": "#/components/schemas/EventDetail"}, {"$ref": "#/components/schemas/LessonDetail"}, {"type": "null"}], "title": "Headline"}, "secondary": {"items": {"anyOf": [{"$ref": "#/components/schemas/EventSummary"}, {"$ref": "#/components/schemas/LessonSummary"}]}, "type": "array", "title": "Secondary", "default": []}, "observance": {"anyOf": [{"$ref": "#/components/schemas/ObservanceRef"}, {"type": "null"}]}}, "additionalProperties": false, "type": "object", "required": ["today"], "title": "TodayResponse", "description": "The full ``/api/v1/today`` payload."}, "TokenPair": {"properties": {"accessToken": {"type": "string", "title": "Accesstoken"}, "refreshToken": {"type": "string", "title": "Refreshtoken"}, "accessExpiresAt": {"type": "string", "format": "date-time", "title": "Accessexpiresat"}, "refreshExpiresAt": {"type": "string", "format": "date-time", "title": "Refreshexpiresat"}, "user": {"$ref": "#/components/schemas/UserPublic"}}, "additionalProperties": false, "type": "object", "required": ["accessToken", "refreshToken", "accessExpiresAt", "refreshExpiresAt", "user"], "title": "TokenPair", "description": "Issued on signup, login, and refresh."}, "UserPublic": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "email": {"type": "string", "format": "email", "title": "Email"}, "displayName": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Displayname"}, "createdAt": {"type": "string", "format": "date-time", "title": "Createdat"}}, "additionalProperties": false, "type": "object", "required": ["id", "email", "displayName", "createdAt"], "title": "UserPublic", "description": "Public account view \u2014 never includes the password hash."}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}